+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix variable resolving not running in client instance on event chain vars
 - Don't crash on `pvp.multiActive` access when server is off
 - Don't crash on `player.username` access when server is off
+- Fix client guis not receiving var change updates
 
 ## [0.7.6] 2026-03-11